Работа с классом races вынесена в отдельный файл
